PCANet parameters set to best ones so far. Test set accuracy is about 72%. Pyramid still not touched.

+++ pcanet_aerial/Utils/im2col_general.m
@@ -9,11 +9,11 @@ z = size(InImg,3);
 im = cell(z,1);
 if NumInput == 2
     for i = 1:z
-        im{i} = im2colstep(double(InImg(:,:,i)),patchsize12)';
+        im{i} = im2colstep(InImg(:,:,i),patchsize12)';
     end
 else
     for i = 1:z
-        im{i} = im2colstep(double(InImg(:,:,i)),patchsize12,varargin{3})';
+        im{i} = im2colstep(InImg(:,:,i),patchsize12,varargin{3})';
     end
 end
 im = [im{:}]';
